Move downloadChromium to the import folder
